Add an error log before raise UnparsedFlagAccessError in Python 2.
This is to make hasattr callers be more aware of the issue, since hasattr silences the exception and returns False in Python 2.

Also make the message more clear on what the flag name is, otherwise if the flag is called "metadata" and the message "Trying to access flag metadata before flags were parsed." is confusing.

PiperOrigin-RevId: 170214560