Allow unicode strings in flag help when used with --helpfull.
Previously, the full help output was gotten by calling str(FLAGS), which did
two things: invoke FLAGS.__str__, which returned a unicode object, and then
make str() try to coerce the unicode object to a byte string, which only worked
if the default encoding and text were compatible.

Rather than coerce the object to a string, just call the get_help() method
directly (which is what FLAGS.__str__ invokes). This avoids any coercion
behaviors.

Whether the unicode text can be printed is still subject to the output stream's
encoding, but thats outside absl's purview: users should configure their
environment to be compatible with unicode as appropriate.

NOTE: One oddity of unicode support with flags is that the default value
for a flag doesn't get printed quite correctly due to repr() being called
on the unicode value: this results in an escaped unicode string (u'\uxxx')
up in the help output instead of the value as expected to be passed on the
command line. That's pre-existing, doesn't break anything, and out of scope.

 * Also improves the error messages of failed app_test helper invocations.

PiperOrigin-RevId: 198766828