Create `app.call_after_init(callback)`.
This allows users to register callbacks which are guaranteed to run after
`app._run_init` has finished and before app.run's main function is invoked.
This is most useful to guarantee setup/init of something is performed when
a library is used under absl.

These may be useful to register callbacks as part of modules being imported
(which usually happens before `app.run`) or in library code that may run in
a context where ABSL's app is not guaranteed to be used by the end-user
(this would also assume that never running the callback is acceptable in that
case).

PiperOrigin-RevId: 205123575