Print test execution status before and after each method.
This helps separate the output (e.g. logging) from the various test methods, making debugging easier when we need to find the logs produced by a particular method.

The new TestResult is a subclass of Python's default TextTestResult, so it passes verbosity=0 to super.__init__ to disable the superclass's per-test output, replacing it completely with its own. Quiet-mode (-q) disables per-test output, just as before.

PiperOrigin-RevId: 207751492