Allow enum.Enum classes to be used directly with flags.
This adds a new flags define, `enum_class`, which is like `enum`, but uses
an `enum.Enum` class as the source of valid values. The resulting parsed
value is the corresponding Enum value.

Using enum_class is recommended over plain enum.

Python 2.7 note: The `enum` module is not required for absl in general, but
is required to use `enum_class`, or for absl tests to fully pass. If the
`enum` module isn't available, then absl flags can still be used, just not
the `enum_class` function. The enum34 package on PyPI is the suggested
package to use for enum support, and is the one absl assumes is used under
Python 2.7.

Python 2.7 and Bazel note: Because of how Bazel modifies PYTHONPATH, absl has
to do do some special sys.path import manipulation to make the enum module
work. In short, third party deps go before the stdlib on sys.path, so enum34's
enum module shadows Python 3's enum module, which causes problems because
enum34 is a subset of later Python 3's enum. So some sys.path futzing is done
to account for this under Bazel. This only happens if 'import enum' initially
fails, and only under Python 2.7.

PiperOrigin-RevId: 212301566