Backport renamed assertRegex names so Py2 can use them.
Several regex asserts were renamed in Python 3 and don't exist in Python 2.
This makes it harder to write code using the Python 3 APIs, since the code is
stuck using the old names and generates warnings in the test output.

The following names are added when run under py2: assertRegex, assertNotRegex,
and assertRaisesRegex

PiperOrigin-RevId: 218389108