Allow defining multi fields with any Iterable, not just lists.
This was motivated to allow tuples as input, but was expanded to allow any
iterable. Strings and non-iterables are still special cased and converted
to a single-element list. A copy of the input iterable is made so that
the flag has sole ownership over the set of values.

NOTE: Custom MultiFlag behavior change: The items of the iterable, rather than
the iterable itself, are now passed onto the underlying self.parser object.
Custom flags using DEFINE_multi or MultiFlag should update their custom flag
code as appropriate.

NOTE: Heavily modified from original PR to adjust for various lint/style
errors and from the auto-formatter.

Resolves #78
Closes #80

PiperOrigin-RevId: 226230348