Clarify the docstring for mark_flags_as_mutual_exclusive
The docstring has a disclaimer clarifying that this specifically checks whether a flag is "set" by checking if the value is not None. That's a good indication that it can gain a lot of clarity by avoiding the use of the word "set" elsewhere in the docstring, instead using "is None" or "is not None" specifically. The disclaimer can also do a little more to call out the fact that default values are not distinguished from explicit values by this validator.

PiperOrigin-RevId: 227527623