Change "should be specified" errors to "value other than None"
Previously, the error message for mark_flags_as_mutual_exclusive said "specified", but "specified" sounds like "included in the command-line invocation. That's not necessarily equivalent, in particular because flags can have default values other than None.

mark_flags_required used a similar wording in the error message, but generated a warning when used with a flag with a default value other than None.

Instead, make both generate a warning and clarify the error message for both to "should have a value other than None", since neither validator checks that a flag value was specified in the command-line invocation. The error message should still be clear in cases where the warning is suppressed or overlooked.

Also adds a test for the existing warning behavior of mark_flag_as_required.

PiperOrigin-RevId: 228510310