Make absl tests pass under new Bazel Python version transition logic.
This makes the absl tests pass when
--incompatible_allow_python_version_transitions=true is set.

Many tests use subprocesses to verify behavior and were relying on
how the subprocess's Python version would match the test's Python
version. Since this stickiness is going away, to keep the test and
subprocess using the same Python version, the binaries have to be
defined twice, one for each Python version, and a select()
used with the test to select the matchiing binary.

NOTE: Under bazel and Python 3, --python_version=PY3 flag must be
specified so that bazel knows how to properly resolve selects.
PiperOrigin-RevId: 243099319