Do not require objects passed to assertTotallyOrdered to implement __hash__
If two objects a and b implement __hash__ and compare equal, hash(a) must equal
hash(b). But it's perfectly valid to define __eq__ (and other comparison
operators) but make the object unhashable. So long as a class properly declares itself unhashable (setting __hash__ = None), assertTotallyOrdered should not attempt to hash objects of the class.

PiperOrigin-RevId: 246150798