Make `TempDir.create_file` chmod read-only files to be writable to initialize them.
The file path passed to `TempDir.create_file()` may already exist when a file
path is given. If it already exists, its file permissions may not mean it
is writable (prior test or code-under-test logic may have made it read-only),
which will break the truncate-on-create initialization logic.

Fixed by checking if it exists and is writable and chmod'ing it appropriately.
This is only done if a file name is given (random file names shouldn't exist,
by definition).

PiperOrigin-RevId: 272042456
Change-Id: I225c3d2ee6e76826585ebf2253905edce67510d0