Record the seed used for test order randomization in test.xml.
Currently, if a user requests randomization of test case execution order, the
only place this is recorded is in a logging statement during test execution.

If a user is both (1) randomizing test case execution order and (2) using the
xml_reporter to produce a test summary, this CL adds a new property to the
top-level testsuites entry including the seed used for randomizing the order
of test cases.

Along the way, I made two incidental changes:

* I renamed the testLoader's `_seed` attribute to `_randomization_order_seed`
  for clarity, as there were at least two seeds already floating around in
  this file (the other being related to `--test_random_seed`).

* I cleaned up some dead code in the test, and commandeered an otherwise-dead
  method (`_run_test_and_get_xml`) for use in a test of the new XML logging.

PiperOrigin-RevId: 292623509
Change-Id: I53da355d4a1c5a11b5e6e8622745fa466840e0ce