When sharding tests, use the test order provided by the TestLoader.
Test sharding works by sorting the list of all tests to execute, and having
shard i take every ith test. (This has the appealing property that each shard
can determine which tests to run independently).

However, if the TestLoader returns the test cases in a different order, we
want to *ignore* that order for selecting tests to execute, but *respect* that
order for the actual execution of the tests, eg in the case of randomized test
ordering.

This change tweaks the sharding implementation to respect the test case order
returned by the TestLoader, and adds a new test.

I also cleaned up the use of `assertEquals` in favor of `assertEqual` while I
was here.

PiperOrigin-RevId: 300132733
Change-Id: Iab86aeb42163c40c24a60af2e779499d0696497d