Fix aliases from marking the aliased flag as present when the alias is defined.
This fixes a bug where the mere act of defining an alias flag
would increment the aliased flag's `Flag.present` counter, thus
incorrectly indicating the aliased flag was always present on
the command line.

This was happening because of the combination of the alias flag's
custom flag parser and the initialization logic of `Flag`.
Essentially, the alias used a custom parser which called the
aliased flag's `Flag.parse` method, and the initialization logic
of `Flag` calls the parser to convert the default value. Thus,
when an alias was defined, the aliased flag's `parse()` method
was called, which increments the `present` counter.

Fixed by making the alias `Flag` subclass avoid calling the parser
during initialization on the default value; the aliased flag
already parsed it.

Also fixes several bugs with aliases:

* Aliasing a multi flag no longer causes the default value to
  be appended to after parsing.
* The alias default value now matches the aliased default value.
* The alias present counter correctly increments after parsing.
  (The aliased present counter still incorrectly increments, but
  that is a pre-existing bug).

PiperOrigin-RevId: 307842972
Change-Id: If36672d4b4b2b261de6c87e7ccf3e595241522b8