The generic parameter of FlagHolder now represents the type of value().
(a) For a flag with default None, define_* will return a FlagHolder[Optional[T]]
(b) For a flag with default != None, define_* will return a FlagHolder[T]

When we are unable to express "default != None", we always return FlagHolder[Optional[T]]

This assumes that the type of default() and value() will be the same.

This assumption holds, till we want the type system to understand that a "required" flag with a None default cannot return a None value.

Changes:
- When exact types for default is known, add @overloads for the DEFINE_* methods for the None case and remaining cases.
- When exact types for default is now known, change return type to FlagHolder[Optional[T]] -- which has same semantics of what FlagHolder[T] used to have.
PiperOrigin-RevId: 315519341
Change-Id: I88dc63b3c5271e12f3274b7246eb602b91b84caf