*Really* ignore errors in `_rmtree_ignore_errors`.
I encountered this in a Windows-specific `OSError` because I made the file to be deleted unwritable, and I guess in Unix you can delete unwritable files, but not Windows. Unfortunately, Windows isn't really a supported platform AIUI, so-- the test will go un-exercised unless I can reproduce on unixy OSes somehow. It turns out that on Unix, too, a file can be made un-deletable, but it's far more obscure -- the directory needs to be missing the write bit. (And to hit the right branch, where absltest thinks it's a file, it must still have the execute bit.)

Of course, the file *should* be deletable, and this is a bug in my test, which should ideally clean up after itself properly. But it says "ignore_errors" right there!

NOTE: this was not actually tested on windows, although the same circumstance should be replicated. YMMV. (I'm kind of wondering right now if the test itself is cross-platform, or if that `os.chmod(dirname, IEXEC)` will throw.)
PiperOrigin-RevId: 329339802
Change-Id: Ib1ec36d754c34a5a10520b8b105269e3c4681fd6