Adds `parameterized.product`: generates the Cartesian product of tests from multiple parameters.
This makes it easier to take a test grid of parameters and generate tests for
each combination. While the usual `parameters` function _can_ do this, it
usually can't be fit into the decorator expression, and the extra logic
to produce the product obscures the intention.

Example usage:

```
class TestModuloExample(parameterized.TestCase):
  @parameterized.product(
      num=[0, 10, 20],
      modulo=[2, 4],
      expected=[0]
  )
  def testModuloResult(self, num, modulo, expected):
    self.assertEqual(expected, num % modulo)
```

PiperOrigin-RevId: 345657677
Change-Id: I7a3dfc868dba291a48c9d3b0e9b77921a3db494f