Update parameterized.CoopTestCase to work with python3 metaclasses.
By setting and accessing __metaclass__, we were not achieving the goals of this
function, which is to make it possible to have both the functionality of
paramterized.TestCase and another TestCase instance with its own metaclass.

Because there are client classes that are using CoopTestCase to combine parameterized.TestCase with another class that isn't using a metaclass we need to support that use-case. We do it by returning a simple multiple inheritance subclass without a combined metaclass, and emit a warning telling the author they don't need to use this method.

PiperOrigin-RevId: 548732415