Pass the shell command to /bin/sh as stdin rather than as an argument
If a command is nearing the maximum command line length of the kernel, this can save us for a bit. The entire command passed in as a string to Bash may exceed the limit whereas the individual process invocations within the passed command would not exceed the limit. We let such situations pass.

While we're here, modernize to use `subprocess.run`, it's equivalent to the prior `Popen`+`communicate`

PiperOrigin-RevId: 683147244